requirements.txt
gitdb==4.0.10
GitPython==3.1.29
PyYAML==6.0
smmap==5.0.0
